requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
greenlet==3.0.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
playwright==1.40.0
pluggy==1.6.0
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.1.0
pytest-playwright==0.4.3
pytest-html==4.1.1
python-slugify==8.0.4
requests==2.31.0
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
